Generate unique classNames
This update (super) enhances Fancy to with the CSS Modules-like feature of
parsing and generating unique classNames.

This is done by tapping into `stylis` with a custom plugin to identify
and tokenize selector names. Several parsing steps have been adding to extract
those classNames and create (short) uuid variants, which are then:

1. rendered (with the original rules) into the `<style>` tag
2. passed into the `ComposedComponent` as a `styles` prop for use

The previous `StyleSheet` factory was converted into a simple Class for
ease of understanding.