FrameManager: Fix IE/Edge frame rendering
This update fixes the iFrame `SCRIPT70` error that occurs in IE/Edge
when an iFrame unmounts/remounts.

The solution was to wrap the offending code in a `try/catch`.

The code in question is designed for performance, specifically to check
for cached instances of Emotion, and serves that instance instead of
creating a brand new one.

The caching mechanism will work for all browsers excluding IE/Edge, which
unfortunately, will create new emotion instances every time as unmounted
frames cannot be properly references.