FrameManager: Fix IE/Edge frame rendering (#33)
## FrameManager: Fix IE/Edge frame rendering

![screen recording 2018-11-06 at 05 24 pm](https://user-images.githubusercontent.com/2322354/48097846-fb534200-e1e8-11e8-9d2c-c48048f01392.gif)

(GIF: Demo'ing Frame 1 being unmounted (destroy) and remounted in Edge. No errors in console. Stays are working)

This update fixes the iFrame `SCRIPT70` error that occurs in IE/Edge
when an iFrame unmounts/remounts.

The solution was to wrap the offending code in a `try/catch`.

The code in question is designed for performance, specifically to check
for cached instances of Emotion, and serves that instance instead of
creating a brand new one.

The caching mechanism will work for all browsers excluding IE/Edge, which
unfortunately, will create new emotion instances every time as unmounted
frames cannot be properly references.

Tested on Edge, IE, and Chrome