Automatically add data-cy and displayName classNames
This update enhances the `styled()` factory to provide the component with
`data-cy` and a `css-j21lkjf-COMPONENT_NAME` className, if applicable.

`data-cy` will be added if the Component has a `displayName` or a `name`.
HTML Tags like `span` or `div` will not receive `data-cy`.

The displayName className will be added to all components. This is purely
for identification/debugging purposes. No actual styles will be associated
with these classNames.